+++ packages/transaction-pay-controller/src/constants.ts
@@ -1,9 +1,7 @@
 import type { Hex } from '@metamask/utils';
 
 export const CONTROLLER_NAME = 'TransactionPayController';
-export const CHAIN_ID_MAINNET = '0x1' as Hex;
 export const CHAIN_ID_ARBITRUM = '0xa4b1' as Hex;
-export const CHAIN_ID_LINEA = '0xe708' as Hex;
 export const CHAIN_ID_POLYGON = '0x89' as Hex;
 export const CHAIN_ID_HYPERCORE = '0x539' as Hex;
 
@@ -16,18 +14,22 @@ export const ARBITRUM_USDC_ADDRESS =
 export const POLYGON_USDCE_ADDRESS =
   '0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174' as Hex;
 
-export const HYPERCORE_USDC_ADDRESS =
-  '0x00000000000000000000000000000000' as Hex;
-
-export const MUSD_TOKEN_ADDRESS =
-  '0xaca92e438df0b2401ff60da7e4337b687a2435da' as Hex;
-
 export const STABLECOINS: Record<Hex, Hex[]> = {
-  [CHAIN_ID_MAINNET]: [MUSD_TOKEN_ADDRESS],
+  // Mainnet
+  '0x1': [
+    '0xaca92e438df0b2401ff60da7e4337b687a2435da', // MUSD
+    '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48', // USDC
+    '0xdac17f958d2ee523a2206206994597c13d831ec7', // USDT
+  ],
   [CHAIN_ID_ARBITRUM]: [ARBITRUM_USDC_ADDRESS.toLowerCase() as Hex],
-  [CHAIN_ID_LINEA]: [MUSD_TOKEN_ADDRESS],
+  // Linea
+  '0xe708': [
+    '0xaca92e438df0b2401ff60da7e4337b687a2435da', // MUSD
+    '0x176211869ca2b568f2a7d4ee941e073a821ee1ff', // USDC
+    '0xa219439258ca9da29e9cc4ce5596924745e12b93', // USDT
+  ],
   [CHAIN_ID_POLYGON]: [POLYGON_USDCE_ADDRESS.toLowerCase() as Hex],
-  [CHAIN_ID_HYPERCORE]: [HYPERCORE_USDC_ADDRESS],
+  [CHAIN_ID_HYPERCORE]: ['0x00000000000000000000000000000000'], // USDC
 };
 
 export enum TransactionPayStrategy {