+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -23,8 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **BREAKING:** Add subsidized fee to Relay quote target amount if `isMaxAmount`.
   - Remove `human` and `raw` from `targetAmount` on `TransactionPayQuote` and `TransactionPayTotals`.
-  - Use `amountFormatted` as USD value for Relay quote target amount when target token is a stablecoin.
-  - Add MUSD and Hypercore USDC to stablecoins.
+  - Use `amountFormatted` as USD value for Relay quote target amount and subsidized fee when token is a stablecoin.
+  - Add MUSD, USDC, USDT, and Hypercore USDC to stablecoins.
 - Bump `@metamask/bridge-controller` from `^65.3.0` to `^66.1.1` ([#7862](https://github.com/MetaMask/core/pull/7862), [#7897](https://github.com/MetaMask/core/pull/7897), [#7910](https://github.com/MetaMask/core/pull/7910))
 - Bump `@metamask/transaction-controller` from `^62.14.0` to `^62.17.0` ([#7854](https://github.com/MetaMask/core/pull/7854), [#7872](https://github.com/MetaMask/core/pull/7872), [#7897](https://github.com/MetaMask/core/pull/7897))
 - Bump `@metamask/assets-controllers` from `^99.2.0` to `^99.3.2` ([#7855](https://github.com/MetaMask/core/pull/7855), [#7860](https://github.com/MetaMask/core/pull/7860)), ([#7897](https://github.com/MetaMask/core/pull/7897))